third_party/WebKit/LayoutTests/html/dialog/dialog-enabled.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<dialog id="mydialog"></dialog>
<script>
description("The DIALOG element should be recognized if the dialog feature is on.");
dialog = document.getElementById('mydialog');
shouldBeTrue("typeof dialog.show == 'function'");
</script>
</body>
</html>
